splitting up protocol and factory from server
